moves header and footer to screens folder
