+++ src/screens/Root.js
@@ -1,7 +1,7 @@
 import React, { Component } from 'react';
-import Header from '../components/Header';
+import Header from './Header';
 import Home from '../components/Home';
-import Footer from '../components/Footer';
+import Footer from './Footer';
 
 
 class ScreenRoot extends Component {
